code review changes - file structure reorganised
