+++ executor/scriptExecutor_test.go
@@ -1,4 +1,4 @@
-package cicdStageExecutor
+package executor
 
 import (
 	"fmt"
